refactor: ListActivity Composable 함수 파일 분리
