+++ app/src/main/java/woowacourse/payments/list/CardListTopBar.kt
@@ -1,0 +1,21 @@
+package woowacourse.payments.list
+
+import androidx.compose.foundation.layout.RowScope
+import androidx.compose.material3.CenterAlignedTopAppBar
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun CardListTopBar(
+    modifier: Modifier = Modifier,
+    actions: @Composable (RowScope.() -> Unit) = {},
+) {
+    CenterAlignedTopAppBar(
+        title = { Text(text = "Payments") },
+        actions = actions,
+        modifier = modifier,
+    )
+}
